+++ src/bitmask/gaps.rs
